+++ packages/react-native/src/Authenticator/index.ts
@@ -1,4 +1,17 @@
 export { default as Authenticator } from './Authenticator';
+export { ContainerProps } from './common';
+export {
+  ConfirmResetPasswordProps,
+  ConfirmSignInProps,
+  ConfirmSignUpProps,
+  ConfirmVerifyUserProps,
+  ForceNewPasswordProps,
+  ForgotPasswordProps,
+  SetupTotpProps,
+  SignInProps,
+  SignUpProps,
+  VerifyUserProps,
+} from './Defaults';
 export { AuthenticatorProps, WithAuthenticatorOptions } from './types';
 export { useAuthenticator, UseAuthenticator } from './useAuthenticator';
 export { default as withAuthenticator } from './withAuthenticator';