+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @aws-amplify/ui-react-native
 
+## 2.4.4
+
+### Patch Changes
+
+- [#6383](https://github.com/aws-amplify/amplify-ui/pull/6383) [`f8319b263cc2df05e12aca672e06755ff8ffcba6`](https://github.com/aws-amplify/amplify-ui/commit/f8319b263cc2df05e12aca672e06755ff8ffcba6) Thanks [@jjarvisp](https://github.com/jjarvisp)! - chore(react-native): update react-native-safe-area-context peer dep
+
+- Updated dependencies [[`e0fcb4322ab8af7e0699da577dc5fbcf9f7e56cc`](https://github.com/aws-amplify/amplify-ui/commit/e0fcb4322ab8af7e0699da577dc5fbcf9f7e56cc)]:
+  - @aws-amplify/ui@6.9.1
+  - @aws-amplify/ui-react-core@3.3.4
+  - @aws-amplify/ui-react-core-notifications@2.2.4
+
 ## 2.4.3
 
 ### Patch Changes